+++ src/Oro/Bundle/ApiBundle/Processor/Get/GetContext.php
@@ -4,6 +4,9 @@ namespace Oro\Bundle\ApiBundle\Processor\Get;
 
 use Oro\Bundle\ApiBundle\Processor\SingleItemContext;
 
+/**
+ * The execution context for processors for "get" action.
+ */
 class GetContext extends SingleItemContext
 {
 }