+++ src/Oro/Bundle/ApiBundle/Processor/Update/UpdateContext.php
@@ -6,6 +6,9 @@ use Oro\Bundle\ApiBundle\Processor\FormContext;
 use Oro\Bundle\ApiBundle\Processor\FormContextTrait;
 use Oro\Bundle\ApiBundle\Processor\SingleItemContext;
 
+/**
+ * The execution context for processors for "update" action.
+ */
 class UpdateContext extends SingleItemContext implements FormContext
 {
     use FormContextTrait;