+++ src/Oro/Bundle/ApiBundle/Provider/ConfigBag.php
@@ -2,13 +2,14 @@
 
 namespace Oro\Bundle\ApiBundle\Provider;
 
-use Oro\Bundle\ApiBundle\Request\Version;
-
 /**
  * A storage for configuration of all registered Data API resources.
  */
 class ConfigBag implements ConfigBagInterface
 {
+    private const ENTITIES  = 'entities';
+    private const RELATIONS = 'relations';
+
     /** @var array */
     private $config;
 
@@ -25,7 +26,7 @@ class ConfigBag implements ConfigBagInterface
      */
     public function getClassNames(string $version): array
     {
-        return array_keys($this->findConfigs('entities', $version));
+        return \array_keys($this->findConfigs(self::ENTITIES, $version));
     }
 
     /**
@@ -33,7 +34,7 @@ class ConfigBag implements ConfigBagInterface
      */
     public function getConfig(string $className, string $version): ?array
     {
-        return $this->findConfig('entities', $className, $version);
+        return $this->findConfig(self::ENTITIES, $className, $version);
     }
 
     /**
@@ -41,7 +42,7 @@ class ConfigBag implements ConfigBagInterface
      */
     public function getRelationConfig(string $className, string $version): ?array
     {
-        return $this->findConfig('relations', $className, $version);
+        return $this->findConfig(self::RELATIONS, $className, $version);
     }
 
     /**
@@ -55,11 +56,8 @@ class ConfigBag implements ConfigBagInterface
         if (!isset($this->config[$section])) {
             return [];
         }
-        $result = $this->config[$section];
 
-        // @todo: API version is not supported for now. Implement filtering by the version here
-
-        return $result;
+        return $this->config[$section];
     }
 
     /**
@@ -77,35 +75,7 @@ class ConfigBag implements ConfigBagInterface
             // no config for the requested class
             return null;
         }
-        $result = $this->config[$section][$className];
-        /* @todo: API version is not supported for now
-        // normalize the version if needed
-        if ($version === Version::LATEST) {
-            $version = null;
-        }
-        if (null !== $version && isset($this->config[$section][$className][$version])) {
-            // found config for exactly requested version
-            return $this->config[$section][$className][$version];
-        }
 
-        $result        = null;
-        $resultVersion = null;
-        foreach ($this->config[$section][$className] as $configVersion => $config) {
-            if (null !== $version && version_compare($configVersion, $version) > 0) {
-                // skip current config because its version is greater that the requested version
-                continue;
-            }
-            if (null === $resultVersion || version_compare($configVersion, $resultVersion) > 0) {
-                $resultVersion = $configVersion;
-                $result        = $config;
-            }
-        }
-
-        if (null !== $result && empty($result)) {
-            $result = null;
-        }
-        */
-
-        return $result;
+        return $this->config[$section][$className];
     }
 }