+++ src/Oro/Bundle/ApiBundle/Util/OptimizeJoinsDecisionMaker.php
@@ -6,7 +6,7 @@ use Doctrine\Common\Collections\Expr\Comparison;
 
 /**
  * Default implementation of the decision maker to decide whether a specific comparison expression
- * from WHERE clause of a query allows or dissalows convertation of a related LEFT join to INNER join.
+ * from WHERE clause of a query allows or disallows conversion of a related LEFT join to INNER join.
  */
 class OptimizeJoinsDecisionMaker implements OptimizeJoinsDecisionMakerInterface
 {