+++ src/Oro/Bundle/TestFrameworkBundle/Behat/Driver/OroSelenium2Driver.php
@@ -3,6 +3,7 @@
 namespace Oro\Bundle\TestFrameworkBundle\Behat\Driver;
 
 use Behat\Mink\Driver\Selenium2Driver;
+use Behat\Mink\Element\NodeElement;
 use Behat\Mink\Selector\Xpath\Escaper;
 use Behat\Mink\Selector\Xpath\Manipulator;
 use Oro\Bundle\TestFrameworkBundle\Behat\Context\AssertTrait;
@@ -392,5 +393,35 @@ var node = {{ELEMENT}};
 triggerEvent(node, '$eventName');
 JS;
         $this->executeJsOnXpath($xpath, $script);
+    }
+
+    /**
+     * @param NodeElement $element
+     */
+    public function switchToIFrameByElement(NodeElement $element)
+    {
+        $id = $element->getAttribute('id');
+
+        if ($id === null) {
+            $elementXpath = $element->getXpath();
+            $id = sprintf('iframe-%s', md5($elementXpath));
+
+            $function = <<<JS
+(function(){
+    var iframeElement = document.evaluate(
+        "{$elementXpath}",
+        document,
+        null,
+        XPathResult.FIRST_ORDERED_NODE_TYPE,
+        null
+    ).singleNodeValue;
+    iframeElement.id = "{$id}";
+})()
+JS;
+
+            $this->executeScript($function);
+        }
+
+        parent::switchToIFrame($id);
     }
 }