+++ src/Oro/Component/ConfigExpression/Tests/Unit/Condition/BlankTest.php
@@ -54,6 +54,11 @@ class BlankTest extends \PHPUnit\Framework\TestCase
                 'options'        => [new PropertyPath('foo')],
                 'context'        => [],
                 'expectedResult' => true
+            ],
+            'empty_array'         => [
+                'options'        => [new PropertyPath('foo')],
+                'context'        => ['foo' => []],
+                'expectedResult' => true
             ]
         ];
     }